+ (wip) SeekPreview on progress bar hover, PoC
This is getting exciting, accidentally converging the design towards a video player-like interface, this is getting less stupid.

I'm not passionate about the traditional filmstrip-like UI with a row of scrollable thumbnails. It's the safest bet in terms of implementing a useful feature, but the thing is, we already see a grid of thumbnails outside the Gallery modal, why bother introducing them back into the "immersive mode" view, except now in one dimension instead of two?